refactor : Study.interface.ts 파일 분리
- 필요없는 코드 제거
- 주석 보완
- 기타 자잘한 리팩토링

#154
