+++ src/pages/Study/Study.interface.ts
@@ -1,0 +1,6 @@
+import { StudyListInfo } from '@api/dto';
+
+export interface ModalInfo {
+  mode: 'add' | 'modify';
+  selectedStudy?: StudyListInfo;
+}
